# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIDB — API POS",
        "description": "API RESTful para el punto de venta (POS) del Sistema Integrado de Distribución y Bodega.<br><br>**Cómo autenticarse:**<br>1. Ejecuta **POST /login** con tu email y password.<br>2. Copia el `token` de la respuesta.<br>3. Haz clic en el botón **Authorize 🔒** (arriba a la derecha).<br>4. Escribe: `Bearer {tu-token}` y confirma.",
        "contact": {
            "name": "Soporte SIDB",
            "email": "soporte@sidb.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/SIDB/public/api",
            "description": "Laragon local"
        }
    ],
    "paths": {
...
        "/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar y buscar clientes activos",
                "description": "Permite buscar por nombre, apellido, DUI, código anterior o teléfono. Devuelve resultados paginados.",
                "operationId": "b910fa16d64cfa03d6400b20f5ffbf0d",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Buscar por nombre, apellido, DUI, código anterior o teléfono",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "7338"
                        }
                    },
                    {
                        "name": "sucursal_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 50
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Paginacion"
                                        },
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Cliente"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "c39fe211a2fc59a9150faf90e77f5568",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellido"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Maria"
                                    },
                                    "apellido": {
                                        "type": "string",
                                        "example": "Hernandez"
                                    },
                                    "dui": {
                                        "type": "string",
                                        "example": "02741470-9"
                                    },
                                    "codigo_anterior": {
                                        "type": "string",
                                        "example": "7338"
                                    },
                                    "telefono_normal": {
                                        "type": "string",
                                        "example": "77856915"
                                    },
                                    "telefono_whatsapp": {
                                        "type": "string",
                                        "example": "77856915"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "maria@email.com"
                                    },
                                    "direccion": {
                                        "type": "string"
                                    },
                                    "departamento": {
                                        "type": "string"
                                    },
                                    "municipio": {
                                        "type": "string"
                                    },
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 13.945
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -87.87
                                    },
                                    "ruta_cobro_id": {
                                        "type": "integer"
                                    },
                                    "foto_casa": {
                                        "description": "Foto de la casa (JPG, PNG, WEBP, máx 4MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/clientes/{id}/ubicacion": {
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar latitud y longitud de un cliente",
                "operationId": "6f6e68210e8ae63aa58615e98dce1dc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitud",
                                    "longitud"
                                ],
                                "properties": {
                                    "latitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 13.945
                                    },
                                    "longitud": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -87.87
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ubicación actualizada"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/clientes/{id}/telefonos": {
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar teléfonos de un cliente",
                "operationId": "f783e7ab7a8e066fbb31d3bc1e1be45c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "telefono_normal": {
                                        "type": "string",
                                        "example": "75123456"
                                    },
                                    "telefono_whatsapp": {
                                        "type": "string",
                                        "example": "75123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Teléfonos actualizados"
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente por ID",
                "operationId": "7a1fb342e3b20be3813e27dbbcfb493e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginacion": {
                "description": "Envoltura de paginación estándar de Laravel",
                "properties": {
                    "current_page": {
                        "description": "Definición central de la especificación OpenAPI para la API POS del SIDB.",
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 5
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 50
                    },
                    "total": {
                        "type": "integer",
                        "example": 230
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "type": "integer",
                        "example": 50
                    }
                },
                "type": "object"
            },
...
            "Cliente": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "dui": {
                        "type": "string",
                        "example": "01234567-8",
                        "nullable": true
                    },
                    "telefono_normal": {
                        "type": "string",
                        "example": "2200-0000",
                        "nullable": true
                    },
                    "telefono_whatsapp": {
                        "type": "string",
                        "example": "7000-0000",
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    },
                    "limite_credito": {
                        "type": "number",
                        "format": "float",
                        "example": 500
                    },
                    "saldo": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Pega aquí el token obtenido en POST /login (sin la palabra Bearer)",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}